+++ nanopub/client.py
@@ -6,6 +6,7 @@ import os
 import random
 import tempfile
 import warnings
+from typing import List, Union, Tuple
 
 import rdflib
 import requests
@@ -18,9 +19,12 @@ from nanopub.publication import Publication
 NANOPUB_GRLC_URLS = ["http://grlc.nanopubs.lod.labs.vu.nl/api/local/local/",
                      "http://130.60.24.146:7881/api/local/local/",
                      "https://openphacts.cs.man.ac.uk/nanopub/grlc/api/local/local/",
-                     "https://grlc.nanopubs.knows.idlab.ugent.be/api/local/local/",
-                     "http://grlc.np.scify.org/api/local/local/",
-                     "http://grlc.np.dumontierlab.com/api/local/local/"]
+                     "http://grlc.np.dumontierlab.com/api/local/local/"
+                     # These servers do currently not support
+                     # find_valid_signed_nanopubs_with_pattern (2020-12-21)
+                     # "https://grlc.nanopubs.knows.idlab.ugent.be/api/local/local/",
+                     # "http://grlc.np.scify.org/api/local/local/",
+                     ]
 NANOPUB_TEST_GRLC_URL = 'http://test-grlc.nanopubs.lod.labs.vu.nl/api/local/local/'
 NANOPUB_FETCH_FORMAT = 'trig'
 NANOPUB_TEST_URL = 'http://test-server.nanopubs.lod.labs.vu.nl/'
@@ -43,7 +47,8 @@ class NanopubClient:
         else:
             self.grlc_urls = NANOPUB_GRLC_URLS
 
-    def find_nanopubs_with_text(self, text: str, max_num_results: int = 1000):
+    def find_nanopubs_with_text(self, text: str, pubkey: str = None,
+                                filter_retracted: bool = True):
         """Text search.
 
         Search the nanopub servers for any nanopubs matching the
@@ -51,10 +56,12 @@ class NanopubClient:
 
         Args:
             text (str): The text to search on
-            max_num_results (int): Maximum number of result, default = 1000
+            pubkey (str): Public key that the matching nanopubs should be signed with
+            filter_retracted (bool): Toggle filtering for publications that are
+                retracted. Default is True, returning only publications that are not retracted.
 
-        Returns:
-            List of dicts depicting matching nanopublications.
+        Yields:
+            dicts depicting matching nanopublications.
             Each dict holds: 'np': the nanopublication uri,
             'date': date of creation of the nanopublication,
             'description': A description of the nanopublication (if found in RDF).
@@ -62,15 +69,17 @@ class NanopubClient:
         """
         if len(text) == 0:
             return []
-
+        endpoint = 'find_signed_nanopubs_with_text'
         params = {'text': text, 'graphpred': '', 'month': '', 'day': '', 'year': ''}
-
-        return self._search(endpoint='find_nanopubs_with_text',
-                            params=params,
-                            max_num_results=max_num_results)
+        if pubkey:
+            params['pubkey'] = pubkey
+        if filter_retracted:
+            endpoint = 'find_valid_signed_nanopubs_with_text'
+        return self._search(endpoint=endpoint,
+                            params=params)
 
     def find_nanopubs_with_pattern(self, subj: str = None, pred: str = None, obj: str = None,
-                                   max_num_results: int = 1000):
+                                   filter_retracted: bool = True, pubkey: str = None):
         """Pattern search.
 
         Search the nanopub servers for any nanopubs matching the given RDF pattern. You can leave
@@ -80,29 +89,35 @@ class NanopubClient:
             subj (str): URI of the subject that you want to match triples on.
             pred (str): URI of the predicate that you want to match triples on.
             obj (str): URI of the object that you want to match triples on.
-            max_num_results (int): Maximum number of result, default = 1000
+            pubkey (str): Public key that the matching nanopubs should be signed with
+            filter_retracted (bool): Toggle filtering for publications that are
+                retracted. Default is True, returning only publications that are not retracted.
 
-        Returns:
-            List of dicts depicting matching nanopublications.
+        Yields:
+            dicts depicting matching nanopublications.
             Each dict holds: 'np': the nanopublication uri,
             'date': date of creation of the nanopublication,
             'description': A description of the nanopublication (if found in RDF).
 
         """
         params = {}
+        endpoint = 'find_signed_nanopubs_with_pattern'
         if subj:
             params['subj'] = subj
         if pred:
             params['pred'] = pred
         if obj:
             params['obj'] = obj
+        if pubkey:
+            params['pubkey'] = pubkey
+        if filter_retracted:
+            endpoint = 'find_valid_signed_nanopubs_with_pattern'
 
-        return self._search(endpoint='find_nanopubs_with_pattern',
-                            params=params,
-                            max_num_results=max_num_results)
+        yield from self._search(endpoint=endpoint,
+                                params=params)
 
-    def find_things(self, type: str, searchterm: str = ' ',
-                    max_num_results=1000):
+    def find_things(self, type: str, searchterm: str = ' ', pubkey: str = None,
+                    filter_retracted: bool = True):
         """Search things (experimental).
 
         Search for any nanopublications that introduce a concept of the given type, that contain
@@ -111,10 +126,12 @@ class NanopubClient:
         Args:
             type (str): A URI denoting the type of the introduced concept
             searchterm (str): The term that you want to search on
-            max_num_results (int): Maximum number of result, default = 1000
+            pubkey (str): Public key that the matching nanopubs should be signed with
+            filter_retracted (bool): Toggle filtering for publications that are
+                retracted. Default is True, returning only publications that are not retracted.
 
-        Returns:
-            List of dicts depicting matching nanopublications.
+        Yields:
+            dicts depicting matching nanopublications.
             Each dict holds: 'np': the nanopublication uri,
             'date': date of creation of the nanopublication,
             'description': A description of the nanopublication (if found in RDF).
@@ -122,37 +139,87 @@ class NanopubClient:
         """
         if searchterm == '':
             raise ValueError(f'Searchterm can not be an empty string: {searchterm}')
-
+        endpoint = 'find_signed_things'
         params = dict()
         params['type'] = type
         params['searchterm'] = searchterm
+        if pubkey:
+            params['pubkey'] = pubkey
+        if filter_retracted:
+            endpoint = 'find_valid_signed_things'
 
-        return self._search(endpoint='find_things',
-                            params=params,
-                            max_num_results=max_num_results, )
+        yield from self._search(endpoint=endpoint,
+                                params=params)
 
-    def _query_grlc(self, params, endpoint):
+    def find_retractions_of(self, source: Union[str, Publication], valid_only=True) -> List[str]:
+        """Find retractions of given URI
+
+        Find all nanopublications that retract a certain nanopublication.
+
+        Args:
+            source (str or nanopub.Publication): URI or Publication object to find retractions for
+            valid_only (bool): Toggle returning only valid retractions, i.e. retractions that are
+                signed with the same public key as the publication they retract. Default is True.
+
+        Returns:
+            List of uris that retract the given URI
+        """
+
+        if isinstance(source, Publication):
+            if source.is_test_publication and not self.use_test_server:
+                warnings.warn('You are trying to find retractions on the production server, '
+                              'whereas this publication lives on the test server')
+            elif not source.is_test_publication and self.use_test_server:
+                warnings.warn('You are trying to find retractions on the test server, '
+                              'whereas this publication lives on the production server')
+            uri = source.source_uri
+        else:
+            uri = source
+
+        if valid_only:
+            source_publication = self.fetch(uri)
+            public_key = source_publication.signed_with_public_key
+            if public_key is None:
+                raise ValueError('The source publication is not signed with a public key')
+        else:
+            public_key = None
+
+        results = self.find_nanopubs_with_pattern(pred=namespaces.NPX.retracts,
+                                                  obj=rdflib.URIRef(uri),
+                                                  pubkey=public_key,
+                                                  filter_retracted=False)
+        return [result['np'] for result in results]
+
+    @staticmethod
+    def _query_grlc(params: dict, endpoint: str, grlc_url: str) -> requests.Response:
+        """Query a specific nanopub server grlc endpoint."""
+        headers = {"Accept": "application/json"}
+        url = grlc_url + endpoint
+        return requests.get(url, params=params, headers=headers)
+
+    def _query_grlc_try_servers(self, params: dict, endpoint: str) -> Tuple[requests.Response, str]:
         """Query the nanopub server grlc endpoint.
 
         Query a nanopub grlc server endpoint (for example: find_text). Try several of the nanopub
         garlic servers.
+
+        Returns:
+            tuple of: r: request response, grlc_url: url of the grlc server used.
         """
-        headers = {"Accept": "application/json"}
         r = None
         random.shuffle(self.grlc_urls)  # To balance load across servers
         for grlc_url in self.grlc_urls:
-            url = grlc_url + endpoint
-            r = requests.get(url, params=params, headers=headers)
+            r = self._query_grlc(params, endpoint, grlc_url)
             if r.status_code == 502:  # Server is likely down
                 warnings.warn(f'Could not get response from {grlc_url}, trying other servers')
             else:
                 r.raise_for_status()  # For non-502 errors we don't want to try other servers
-                return r
+                return r, grlc_url
 
         raise requests.HTTPError(f'Could not get response from any of the nanopub grlc '
                                  f'endpoints, last response: {r.status_code}:{r.reason}')
 
-    def _search(self, endpoint: str, params: dict, max_num_results: int):
+    def _search(self, endpoint: str, params: dict):
         """
         General nanopub server search method. User should use e.g. find_nanopubs_with_text,
         find_things etc.
@@ -160,24 +227,31 @@ class NanopubClient:
         Args:
             endpoint: garlic endpoint to query, for example: find_things
             params: dictionary with parameters for get request
-            max_num_results: Maximum number of results to return
 
         Raises:
             JSONDecodeError: in case response can't be serialized as JSON, this can happen due to a
                 virtuoso error.
         """
-        r = self._query_grlc(params, endpoint)
-        results_json = r.json()
-
-        results_list = results_json['results']['bindings']
-        nanopubs = []
-
-        for result in results_list:
-            nanopubs.append(self._parse_search_result(result))
-            if len(nanopubs) >= max_num_results:
-                break
-
-        return nanopubs
+        has_results = True
+        page_number = 1
+        grlc_url = None
+        while has_results:
+            params['page'] = page_number
+            # First try different servers
+            if grlc_url is None:
+                r, grlc_url = self._query_grlc_try_servers(params, endpoint)
+            # If we have found a grlc server we should use that for further queries (so
+            # pagination works properly)
+            else:
+                r = self._query_grlc(params, endpoint, grlc_url)
+                r.raise_for_status()
+            results = r.json()
+            bindings = results['results']['bindings']
+            if not bindings:
+                has_results = False
+            page_number += page_number
+            for result in bindings:
+                yield self._parse_search_result(result)
 
     @staticmethod
     def _parse_search_result(result: dict):
@@ -297,17 +371,11 @@ class NanopubClient:
                 match the public key in the profile
         """
         publication = self.fetch(uri)
-        their_public_keys = list(publication.pubinfo.objects(rdflib.URIRef(uri + '#sig'),
-                                                             namespaces.NPX.hasPublicKey))
-        if len(their_public_keys) > 0:
-            their_public_key = str(their_public_keys[0])
-            if len(their_public_keys) > 1:
-                warnings.warn(f'Nanopublication is signed with multiple public keys, we will use '
-                              f'this one: {their_public_key}')
-            if their_public_key != profile.get_public_key():
-                raise AssertionError('The public key in your profile does not match the public key'
-                                     'that the publication that you want to retract is signed '
-                                     'with. Use force=True to force retraction anyway.')
+        their_public_key = publication.signed_with_public_key
+        if their_public_key is not None and their_public_key != profile.get_public_key():
+            raise AssertionError('The public key in your profile does not match the public key'
+                                 'that the publication that you want to retract is signed '
+                                 'with. Use force=True to force retraction anyway.')
 
     def retract(self, uri: str, force=False):
         """Retract a nanopublication.