+++ CHANGELOG.md
@@ -4,6 +4,43 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.2.0] - 2020-12-23
+
+### Added
+* Added Zenodo badge to README
+* Pagination of results for search methods of `NanopubClient`
+
+### Changed
+* `nanopub-java` dependency is installed upon installation instead of upon runtime.
+* search methods of `NanopubClient` return iterator instead of list
+
+
+## [1.1.0] - 2020-12-17
+
+### Added
+* `.zenodo.json` for linking to zenodo
+* `pubkey` option to methods of `NanopubClient` that allows searching for publications 
+    signed with the given pubkey. For these methods:
+    - `find_nanopubs_with_text`
+    - `find_nanopubs_with_pattern`
+    - `find_things`
+* `filter_retracted` option to methods of `NanopubClient` that allows searching for publications 
+    that are note retracted. For these methods:
+    - `find_nanopubs_with_text`
+    - `find_nanopubs_with_pattern`
+    - `find_things`
+* `NanopubClient.find_retractions_of` method to search retractions of a given nanopublication.
+* `Publication.signed_with_public_key` property: the public key that the publication was signed with.
+* `Publication.is_test_publication` property: denoting whether this is a publicaion on the test server.
+
+### Changed
+* Improved error message by pointing to documentation instead of Readme upon ProfileErrors
+
+### Fixed
+* Catch FileNotFoundError when profile.yml does not exist, raise ProfileError with useful messageinstead.
+* Fixed broken link to documentation in README.md
+
 ## [1.0.0] - 2020-12-08
 
 NB: All changes before [1.0.0] are collapsed in here (even though there were multiple pre-releases)